# openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Melogold API",
    "version": "1",
    "description": "Sync server of Melogold: favorites, library, playlists, history and playback across a user's devices. The normative contract is docs/API.md; clients branch only on `code` of ErrorResponse (error-codes.json). There is no version in the URL: compatibility comes from `apiVersion`/`minApiVersion` and `features.sync.protocol` of `GET /server/info` (API §1.1).",
    "license": {
      "name": "CC0-1.0",
      "url": "https://creativecommons.org/publicdomain/zero/1.0/"
    }
  },
  "servers": [
    {
      "url": "https://api.melogold.app",
      "description": "The official server. Self-hosted servers serve the same API under their own base URL."
    }
  ],
...
  "paths": {
...
    "/auth/me/devices": {
      "get": {
        "operationId": "listDevices",
        "summary": "The devices of the account",
        "tags": [
          "devices"
        ],
        "description": "The current device first, the others by lastSeenAt descending (API §4.4).\n\nRate limit: 120/min per user (API §1.10).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceListResponse"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized`, `access_token_invalid`, `access_token_expired`, `session_revoked`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` (retryAfterSeconds)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "`internal_error`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "`server_busy` (retryAfterSeconds), `unavailable` (retryAfterSeconds?)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me/devices/{deviceId}": {
      "patch": {
        "operationId": "renameDevice",
        "summary": "Rename a device",
        "tags": [
          "devices"
        ],
        "description": "`name: null` returns to the reported name. A recently added device must send `password` to rename another device (DESIGN §4.8).\n\nRate limit: 120/min per user (API §1.10).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameDeviceRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
            },
            "in": "path",
            "name": "deviceId",
            "required": true,
            "description": "UUID, lowercase only (API §1.6)."
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceDto"
                }
              }
            }
          },
          "400": {
            "description": "`invalid_request` (issues), `invalid_json`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized`, `access_token_invalid`, `access_token_expired`, `session_revoked`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "`invalid_password`, `recent_device_restricted`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "`device_not_found`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "`payload_too_large`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "`unsupported_media_type`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` (retryAfterSeconds), `reauth_throttled` (retryAfterSeconds)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "`internal_error`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "`server_busy` (retryAfterSeconds), `unavailable` (retryAfterSeconds?)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me/devices/{deviceId}/revoke": {
      "post": {
        "operationId": "revokeDevice",
        "summary": "Revoke another device",
        "tags": [
          "devices"
        ],
        "description": "After commit: session.invalidated{device_revoked} to that device, its streams close, devices.updated{device_removed} to the others (API §4.4). Not retried automatically.\n\nRate limit: 120/min per user (API §1.10).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeDeviceRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
            },
            "in": "path",
            "name": "deviceId",
            "required": true,
            "description": "UUID, lowercase only (API §1.6)."
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "`invalid_request` (issues), `invalid_json`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized`, `access_token_invalid`, `access_token_expired`, `session_revoked`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "`invalid_password`, `recent_device_restricted`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "`device_not_found`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "`cannot_revoke_current_device`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "`payload_too_large`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "`unsupported_media_type`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` (retryAfterSeconds), `reauth_throttled` (retryAfterSeconds)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "`internal_error`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "`server_busy` (retryAfterSeconds), `unavailable` (retryAfterSeconds?)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me/devices/revoke-others": {
      "post": {
        "operationId": "revokeOtherDevices",
        "summary": "Sign out every other device",
        "tags": [
          "devices"
        ],
        "description": "The same rules for every target; on 403 no device is removed (API §4.4).\n\nRate limit: 120/min per user (API §1.10).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeOthersRequest"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokeOthersResponse"
                }
              }
            }
          },
          "400": {
            "description": "`invalid_request` (issues), `invalid_json`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized`, `access_token_invalid`, `access_token_expired`, `session_revoked`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "`invalid_password`, `recent_device_restricted`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "`payload_too_large`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "`unsupported_media_type`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` (retryAfterSeconds), `reauth_throttled` (retryAfterSeconds)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "`internal_error`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "`server_busy` (retryAfterSeconds), `unavailable` (retryAfterSeconds?)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token of TokenPair (API §1.7)."
      }
    },
    "schemas": {
...
      "DeviceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
            "description": "UUID, lowercase."
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64,
            "description": "customName ?? reportedName."
          },
          "reportedName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "customName": {
            "nullable": true,
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "platform": {
            "type": "string",
            "pattern": "^[a-z0-9_]{1,16}$"
          },
          "osVersion": {
            "nullable": true,
            "type": "string",
            "maxLength": 64
          },
          "model": {
            "nullable": true,
            "type": "string",
            "maxLength": 64
          },
          "clientVersion": {
            "nullable": true,
            "type": "string",
            "maxLength": 64
          },
          "linkedVia": {
            "type": "string",
            "description": "Known values: register, login, link, recovery. Clients must accept unknown values (API §1.3)."
          },
          "linkedByDeviceId": {
            "nullable": true,
            "type": "string",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
            "description": "UUID, lowercase."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$",
            "description": "UTC, always `YYYY-MM-DDTHH:mm:ss.sssZ` (API §1.5)."
          },
          "lastSeenAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$",
            "description": "UTC, always `YYYY-MM-DDTHH:mm:ss.sssZ` (API §1.5)."
          },
          "lastSyncAt": {
            "nullable": true,
            "type": "string",
            "format": "date-time",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$",
            "description": "UTC, always `YYYY-MM-DDTHH:mm:ss.sssZ` (API §1.5)."
          },
          "recentUntil": {
            "nullable": true,
            "description": "Until this moment the device is new (DESIGN §4.8), otherwise null.",
            "type": "string",
            "format": "date-time",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
          },
          "isCurrent": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "reportedName",
          "customName",
          "platform",
          "osVersion",
          "model",
          "clientVersion",
          "linkedVia",
          "linkedByDeviceId",
          "createdAt",
          "lastSeenAt",
          "lastSyncAt",
          "recentUntil",
          "isCurrent"
        ]
      },
...
      "DeviceListResponse": {
        "type": "object",
        "properties": {
          "devices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DeviceDto"
            },
            "description": "The current device first, the others by lastSeenAt descending."
          },
          "maxDevices": {
            "nullable": true,
            "description": "`null`: no limit.",
            "type": "integer",
            "minimum": 0,
            "maximum": 2147483647,
            "format": "int32"
          }
        },
        "required": [
          "devices",
          "maxDevices"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "format": "int64"
          },
          "error": {
            "type": "string",
            "description": "Same text as `message` (Clementine shape)."
          },
          "message": {
            "type": "string",
            "description": "English text for logs; not for UI or logic. 5xx always carry the generic text of their code."
          },
          "code": {
            "type": "string",
            "description": "The only field clients branch on (API §2.2, `error-codes.json`). Clients must accept unknown codes."
          },
          "retryAfterSeconds": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "format": "int64"
          },
          "issues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            }
          },
          "minLength": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "format": "int64"
          },
          "maxLength": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "format": "int64"
          },
          "deviceLimit": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "format": "int64"
          },
          "deviceCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "format": "int64"
          },
          "minProtocol": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "format": "int64"
          },
          "maxProtocol": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "format": "int64"
          },
          "floorCursor": {
            "type": "string",
            "description": "Opaque cursor (API §1.6)."
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "code"
        ],
        "description": "Every refusal (API §2.1). Details appear only with their codes."
      },
...
      "RenameDeviceRequest": {
        "type": "object",
        "properties": {
          "name": {
            "nullable": true,
            "description": "`null` returns to the reported name.",
            "minLength": 1,
            "maxLength": 64,
            "type": "string"
          },
          "password": {
            "nullable": true,
            "type": "string",
            "minLength": 1,
            "maxLength": 128,
            "description": "1..128 UTF-16 units and at most 512 UTF-8 bytes; NFKC before verification."
          }
        },
        "required": [
          "name"
        ]
      },
...
      "RevokeDeviceRequest": {
        "type": "object",
        "properties": {
          "password": {
            "nullable": true,
            "type": "string",
            "minLength": 1,
            "maxLength": 128,
            "description": "1..128 UTF-16 units and at most 512 UTF-8 bytes; NFKC before verification."
          }
        }
      },
      "RevokeOthersRequest": {
        "type": "object",
        "properties": {
          "password": {
            "nullable": true,
            "type": "string",
            "minLength": 1,
            "maxLength": 128,
            "description": "1..128 UTF-16 units and at most 512 UTF-8 bytes; NFKC before verification."
          }
        }
      },
      "RevokeOthersResponse": {
        "type": "object",
        "properties": {
          "revokedCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "format": "int64"
          }
        },
        "required": [
          "revokedCount"
        ]
      },
...
      "ValidationIssue": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string",
            "description": "Dot path into the request, e.g. \"ops.3.opId\"."
          },
          "code": {
            "type": "string",
            "description": "zod issue code, e.g. invalid_type, too_big, invalid_format."
          }
        },
        "required": [
          "path",
          "code"
        ]
      }
    }
  }
}